Added test case for #9
Also fixed many bugs relating to packing, which hasn't been properly tested

TODO; add an issue mentioning the need for explicit tests to convert a ChunkyFS to a blob and back.

Blob mapping shouldn't be required to be 1-1 (although it would be nice) BUT the Blob to FS mapping should be. All blobs created from FSs should recreate the same FS.